refactor(cliproxy): extract fro-bot workflow analyzer into setup/workflow-analyzer.ts
Move FroBotWorkflowCheck type, REQUIRED_OPENCODE_INPUTS, findFroBotAgentStepBodies,
analyzeFroBotWorkflow, interpretGhContentResult, checkFroBotWorkflow, and
formatWorkflowSnippet into a focused submodule.

setup.ts shrinks from 884 to 758 LOC. Adds 5 new TDD-first tests for the
newly-exported findFroBotAgentStepBodies helper, including the sibling-step
scoping case from PR #125 (720 total, +5).
